Refactor optimize.py into several files
