+++ LICENSE
@@ -1,6 +1,6 @@
                        BSD 3-clause (aka BSD 2.0) License
 
-Copyright 2016-2019 Regents of the University of California and the Authors
+Copyright 2016-2020 Regents of the University of California and the Authors
 
 Authors: Lee-Ping Wang, Chenchen Song
