+++ geometric/logTest.ini
@@ -1,0 +1,22 @@
+[loggers]
+keys=root
+
+[handlers]
+keys=stream_handler
+
+[formatters]
+keys=formatter
+
+[logger_root]
+level=INFO
+handlers=stream_handler
+
+[handler_stream_handler]
+class=geometric.nifty.RawStreamHandler
+level=INFO
+formatter=formatter
+args=(sys.stderr,)
+
+[formatter_formatter]
+format=%(message)s
+#format=%(asctime)s %(name)-12s %(levelname)-8s %(message)s
