+++ geometric/run_json.py
@@ -1,3 +1,38 @@
+"""
+run_json.py: JSON API for geomeTRIC
+
+Copyright 2016-2020 Regents of the University of California and the Authors
+
+Authors: Yudong Qiu, Daniel G. A. Smith, Lee-Ping Wang
+
+Contributors: 
+
+Redistribution and use in source and binary forms, with or without modification,
+are permitted provided that the following conditions are met:
+
+1. Redistributions of source code must retain the above copyright notice,
+this list of conditions and the following disclaimer.
+
+2. Redistributions in binary form must reproduce the above copyright notice,
+this list of conditions and the following disclaimer in the documentation
+and/or other materials provided with the distribution.
+
+3. Neither the name of the copyright holder nor the names of its contributors
+may be used to endorse or promote products derived from this software
+without specific prior written permission.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
+IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
+INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT
+NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+POSSIBILITY OF SUCH DAMAGE.
+"""
+
 #!/usr/bin/env python
 
 import copy
@@ -176,7 +211,7 @@ def geometric_run_json(in_json_dict):
     constraints_string = make_constraints_string(constraints_dict)
     Cons, CVals = None, None
     if constraints_string:
-        Cons, CVals = geometric.optimize.ParseConstraints(M, constraints_string)
+        Cons, CVals = geometric.prepare.parse_constraints(M, constraints_string)
 
     # set up the internal coordinate system
     coordsys = input_opts.get('coordsys', 'tric')
