+++ geometric/tests/test_hessian.py
@@ -10,8 +10,9 @@ from . import addons
 from geometric.internal import *
 
 datad = addons.datad
+test_logger = addons.test_logger
 
-def test_hessian_assort():
+def test_hessian_assort(test_logger):
     M = geometric.molecule.Molecule(os.path.join(datad, 'assort.xyz'))
     coords = M.xyzs[0].flatten() * ang2bohr
     # Build TRIC coordinate system
